fix: update GitHubIssue model to use long for Number and correct PrintService to use Size in push events

+++ Models/GitHubIssue.cs
@@ -5,7 +5,7 @@ namespace GitHubUserActivity.Models;
 public class GitHubIssue
 {
     [JsonPropertyName("number")]
-    public required int Number { get; init; }
+    public required long Number { get; init; }
     
     [JsonPropertyName("pull_request")]
     public object? PullRequest { get; init; }
